+++ app/src/main/scala/net/sailware/resumewizard/resume/Skill.scala
@@ -5,7 +5,7 @@ import net.sailware.resumewizard.jooq.tables.records.ResumeSkillsRecord
 case class Skill(
   val id: Int,
   val name: String,
-  val rating: Short
+  val rating: Int
 )
 
 object Skill:
