+++ app/src/main/scala/net/sailware/resumewizard/resume/wizard/skill/view/model/SkillEntryForm.scala
@@ -3,5 +3,5 @@ package net.sailware.resumewizard.resume.wizard.skill.view.model
 case class SkillEntryForm(
   id: Int,
   name: String,
-  rating: Short
+  rating: Int
 )
