+++ app/src/main/scala/net/sailware/resumewizard/resume/wizard/skill/view/model/SkillNewEntryForm.scala
@@ -2,5 +2,5 @@ package net.sailware.resumewizard.resume.wizard.skill.view.model
 
 case class SkillNewEntryForm(
   val name: String,
-  val rating: Short
+  val rating: Int
 )
